Use liblog_for_runtime_apex instead of liblog liblog_for_runtime_apex is a static variant of liblog which is explicitly marked as available to the runtime APEX. Any static dependency to liblog from inside the runtime APEX is changed from liblog to liblog_for_runtime_apex.
Previously, to support the need for using liblog inside the runtime
APEX, the entire (i.e. both static and shared variants) liblog module
was marked as available to the runtime APEX, although in reality only
the static variant of the library was needed there. This was not only
looking dirty, but also has caused a problem like b/241259844.

To fix this, liblog is separated into two parts. (1) liblog and (2)
liblog_for_runtime_apex. (1) no longer is available to the runtime APEX
and is intended to be depended on in most cases: either from the
non-updatable platform, or from other APEXes. (2) is a static library
which is explicitly marked as available to the runtime APEX and also
visible to certain modules that are included in the runtime APEX.

Bug: 241259844
Test: m and check that liblog depends on stub library of libc
Change-Id: I74166537f660f7dc7c59c3739c584c0443b04e6f
Signed-off-by: escobar1945 <candaroglu197878@gmail.com>